Split up AA rest api into a client
